+++ example/example.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/foundation.dart';
 import 'package:pexels_photos_videos/pexels_media.dart';
-import 'package:pexels_photos_videos/pexels_result.dart';
+import 'package:pexels_photos_videos/photo.dart';
 
 void main() async {
   /// Call [PexelsMedia] and get [PexelsResult]
-  PexelsResult result = await PexelsMedia()
-      .getPexelsPhotos('place your auth key here', 'image Id');
+  Photo photo = await PexelsMedia()
+      .getPhoto('place your auth key here', 'image Id');
   if (kDebugMode) {
-    print(result.photographer);
+    print(photo.photographer);
   }
 }