+++ test/pexels_photos_videos_test.dart
@@ -5,7 +5,7 @@ import 'package:pexels_photos_videos/pexels_media.dart';
 void main() {
   test('pass your own authorization key and image id to run this test', () {
     final pexelsMedia = PexelsMedia();
-    expect(pexelsMedia.getPexelsPhotos('authorizationKey', 'imageId'),
+    expect(pexelsMedia.getPhoto('authorizationKey', 'imageId'),
         'Invalid Authentication');
   });
 }